+++ src/scalar.rs
@@ -13,8 +13,10 @@ use core::fmt;
 use core::iter::{Product, Sum};
 use core::ops::{Add, AddAssign, BitAnd, BitXor, Mul, MulAssign, Neg, Sub, SubAssign};
 use rand_core::{CryptoRng, RngCore};
-use serde::{de::Visitor, Deserialize, Deserializer, Serialize, Serializer};
 use subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption};
+
+#[cfg(feature = "serde_req")]
+use serde::{de::Visitor, Deserialize, Deserializer, Serialize, Serializer};
 
 /// Represents an element of the scalar field $\mathbb{F}_q$ of the BLS12-381 elliptic
 /// curve construction.
@@ -88,6 +90,7 @@ impl ConditionallySelectable for Scalar {
     }
 }
 
+#[cfg(feature = "serde_req")]
 impl Serialize for Scalar {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
@@ -102,6 +105,7 @@ impl Serialize for Scalar {
     }
 }
 
+#[cfg(feature = "serde_req")]
 impl<'de> Deserialize<'de> for Scalar {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
@@ -1374,6 +1378,7 @@ fn test_iter_prod() {
 }
 
 #[test]
+#[cfg(feature = "serde_req")]
 fn serde_bincode_scalar_roundtrip() {
     use bincode;
     let scalar = -Scalar::from(3u64);