+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 24-12-20
+### Changed
+- no-std compatibility for pairings feature
+- isolate serde with `serde_req` feature
+
 ## [0.3.0] - 08-11-20
 ### Changed
 - no-std compatibility